local_ocr_server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pymupdf==1.24.14
pillow==10.4.0
easyocr==1.7.1
numpy==1.26.4
